+++ GVFS/GVFS.FunctionalTests/Tests/GitCommands/GitCommandsTests.cs
@@ -565,7 +565,7 @@ namespace GVFS.FunctionalTests.Tests.GitCommands
         {
             const string TestFileName = "test-file-mode";
             this.CreateFile("#!/bin/bash\n", TestFileName);
-            this.ChangeMode(TestFileName, Convert.ToInt32("755", 8));
+            this.ChangeMode(TestFileName, Convert.ToUInt16("755", 8));
             this.ValidateGitCommand($"add {TestFileName}");
             this.ValidateGitCommand($"ls-files --stage {TestFileName}");
         }
